requirements.txt
ipykernel
python-dotenv
groq
google-colab
together
openai
python-dotenv
streamlit
